% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_shots.R
\name{scrape_shots}
\alias{scrape_shots}
\title{Scrapes the NHL API for Shot Location Data}
\usage{
scrape_shots(start_season = 2016, end_season = 2016)
}
\arguments{
\item{...}{Passed to [ggplot2::geom_point()]}
}
\description{
Scrapes the NHL Live Data for shot location data including coordinates, shot type,
}
\examples{
scrape_shots(start_season=2016, end_season=2016)
}
\keyword{curl}
\keyword{json}
