% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_shots.R
\name{scrape_shots}
\alias{scrape_shots}
\title{Scrapes the NHL API for Shot Location Data}
\usage{
scrape_shots(start_season = 2016, end_season = 2016)
}
\arguments{
\item{start_season}{The starting year of the first season to be scraped. For example, "2016"
for the 2016-2017 season.}

\item{end_season}{The starting year of the last season to be scraped. For example, "2016"
for the 2016-2017 season.}
}
\description{
Scrapes the NHL Live Data for shot location data including coordinates, shot type,
shooting player name and team.
}
\examples{
scrape_shots(start_season=2016, end_season=2016)
}
\keyword{curl}
\keyword{json}
